refactored parsers into functions
