--- pkgs/swiftgen/swift2objc/lib/src/parser/parsers/symbolgraph_json_reader.dart
@@ -1,15 +1,0 @@
-import 'dart:convert';
-import 'dart:io';
-
-import '../_core/utils.dart';
-
-class SymbolgraphJsonReader {
-  final String symbolgraphJsonPath;
-
-  SymbolgraphJsonReader(this.symbolgraphJsonPath);
-
-  JsonMap read() {
-    final jsonStr = File(symbolgraphJsonPath).readAsStringSync();
-    return jsonDecode(jsonStr);
-  }
-}
