+++ pkgs/swiftgen/swift2objc/lib/src/parser/parsers/wire_up_relations.dart
@@ -1,0 +1,17 @@
+import '../../ast/declarations/compounds/class_declaration.dart';
+import '../_core/utils.dart';
+
+void wireUpRelations(DeclarationsMap declarations, JsonMap symbolgraphJson) {
+  final List relations = symbolgraphJson["relationships"];
+  for (final relation in relations) {
+    if (relation["kind"] != "memberOf") continue;
+
+    String methodId = relation["source"];
+    String classId = relation["target"];
+
+    final methodDeclaraion = declarations[methodId] as ClassMethodDeclaration;
+    final classDeclaraion = declarations[classId] as ClassDeclaration;
+
+    classDeclaraion.methods.add(methodDeclaraion);
+  }
+}
